+++ CHANGELOG.md
@@ -123,6 +123,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Rename `vcd` to `cloud-director`
+- Monitor using a podmonitor.
 
 ### Fixed